backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
openai==1.30.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.131.0
imaplib2==3.6
python-multipart==0.0.9
pydantic==2.7.1
pandas==2.2.2
markdown==3.6
websockets==12.0
